requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.1
pydantic-settings==2.1.0
alembic==1.12.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.0
python-dateutil==2.8.2 